Small refactoring (nested repositories discovery) and fix tests for gateway & tracker modules
