+++ tracker/src/main/java/io/voteofconf/tracker/repository/ExpertiseMWRepository.java
@@ -2,7 +2,6 @@ package io.voteofconf.tracker.repository;
 
 import io.voteofconf.common.model.Expertise;
 import io.voteofconf.common.model.User;
-import io.voteofconf.tracker.repository.support.M2MMappingMWRepository;
 import org.springframework.data.r2dbc.core.DatabaseClient;
 import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Flux;
