--- tracker/src/main/java/io/voteofconf/tracker/repository/support/ClientAgreementsAGRepository.java
@@ -1,6 +1,0 @@
-package io.voteofconf.tracker.repository.support;
-
-import org.springframework.data.repository.reactive.ReactiveCrudRepository;
-
-public interface ClientAgreementsAGRepository extends ReactiveCrudRepository<M2MMappingMWRepository.ClientAgreements, Long> {
-}
