--- tracker/src/main/java/io/voteofconf/tracker/repository/support/UserVacancyAGRepository.java
@@ -1,6 +1,0 @@
-package io.voteofconf.tracker.repository.support;
-
-import org.springframework.data.repository.reactive.ReactiveCrudRepository;
-
-public interface UserVacancyAGRepository extends ReactiveCrudRepository<M2MMappingMWRepository.UserVacancy, Long> {
-}
