+++ tracker/src/test/java/io/voteofconf/tracker/configuration/DatasourceConfiguration.java
@@ -6,22 +6,31 @@ import io.voteofconf.tracker.repository.UserAGCrudRepository;
 import org.mockito.Mockito;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.DependsOn;
 import org.springframework.context.annotation.Profile;
+import org.springframework.data.r2dbc.config.AbstractR2dbcConfiguration;
+import org.springframework.data.r2dbc.connectionfactory.R2dbcTransactionManager;
 import org.springframework.data.r2dbc.core.DatabaseClient;
+import org.springframework.data.r2dbc.repository.config.EnableR2dbcRepositories;
+import org.springframework.transaction.ReactiveTransactionManager;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 @Profile("test")
+@EnableR2dbcRepositories(
+        basePackages = "io.voteofconf.tracker.repository",
+        considerNestedRepositories = true)
 @Configuration
-public class DatasourceConfiguration {
+@EnableTransactionManagement
+public class DatasourceConfiguration extends AbstractR2dbcConfiguration {
 
     @Bean
-    public DatabaseClient databaseClient() {
-
-        ConnectionFactory connectionFactory = ConnectionFactories.get("r2dbc:h2:mem:///test?options=DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE");
-        return DatabaseClient.create(connectionFactory);
+    public ConnectionFactory connectionFactory() {
+        return ConnectionFactories.get("r2dbc:h2:mem:///test?options=DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE");
     }
 
     @Bean
-    public UserAGCrudRepository userRepository() {
-        return Mockito.mock(UserAGCrudRepository.class);
+    @DependsOn("connectionFactory")
+    public ReactiveTransactionManager reactiveTransactionManager(ConnectionFactory connectionFactory) {
+        return new R2dbcTransactionManager(connectionFactory);
     }
 }
